Issue #288: CEXC hangs with multibyte characters in Python 3
Python 3's stdin understands unicode and thinks read() should take a number of characters rather than a number of bytes. The protocol, however, is to send a number of bytes in the message header. When there are multibyte characters, the SDK tries to read more bytes than are available and read() blocks.

* When not in Python 2, replace input_files containing sys.stdin with sys.stdin.buffer in dispatch() so we have a byte-oriented stream. The change happens inside the function as many users are explicitly passing sys.stdin as a positional argument.
* Decode data read from ifile after reading, but only when not in Python 2. unicode instances of multibyte UTF-8 characters interact badly with csv.reader(), as it attempts to decode everything as ascii.